Refactor GalleryPage tests
