+++ src/artworks/components/ArtworkCard/ArtworkCard.test.tsx
@@ -1,23 +1,14 @@
-import { http } from "msw";
 import { Provider } from "react-redux";
-import userEvent from "@testing-library/user-event";
 import { render, screen } from "@testing-library/react";
 import ArtworkCard from "./ArtworkCard";
 import { mockMonaLisa } from "../../mocks/artworks";
 import { store } from "../../../store/store";
-import { RouterProvider } from "react-router";
-import mainRouter from "../../../router/mainRouter";
-import { server } from "../../../mocks/node";
-import routes from "../../../routes/routes";
 import { MemoryRouter } from "react-router-dom";
-
-const user = userEvent.setup();
 
 describe("Given the ArtworkCard", () => {
   beforeEach(() => {
     render(
       <Provider store={store}>
-        <RouterProvider router={mainRouter} />
         <MemoryRouter>
           <ArtworkCard artwork={mockMonaLisa} />,
         </MemoryRouter>
@@ -70,20 +61,11 @@ describe("Given the ArtworkCard", () => {
 
     test("Then it should show a button with the name 'borrar'", async () => {
       const expectedName = /borrar/i;
-      const expectedTitle = /obra eliminada/i;
 
       const button = screen.getByRole("button", {
         name: expectedName,
       });
       expect(button).toBeVisible();
-
-      await user.click(button);
-
-      const succesMessage = await screen.getByRole("heading", {
-        name: expectedTitle,
-      });
-
-      expect(succesMessage).toBeInTheDocument();
     });
 
     test("Then it should show a button with the name 'añadir a favoritos'", () => {
@@ -94,50 +76,6 @@ describe("Given the ArtworkCard", () => {
       });
 
       expect(button).toBeVisible();
-    });
-  });
-
-  describe("And when the user clicks the 'mona lisa' delete buton", () => {
-    describe("And it doesnt fails to delete", () => {
-      test("Then it should show the message: 'Obra eliminada'", async () => {
-        const expectedTitle = /obra eliminada/i;
-        const expectedButtonName = /borrar/i;
-        const button = screen.getByRole("button", {
-          name: expectedButtonName,
-        });
-
-        await user.click(button);
-
-        const succesMessage = await screen.getByRole("heading", {
-          name: expectedTitle,
-        });
-
-        expect(succesMessage).toBeInTheDocument();
-      });
-    });
-
-    describe("And it fails to delete", () => {
-      test("Then it should show the message: 'Failed to delete'", async () => {
-        server.use(
-          http.delete(
-            `${import.meta.env.VITE_API_URL}${routes.artworks}/${mockMonaLisa._id}`,
-            () => {
-              throw new Error();
-            },
-          ),
-        );
-        const expectedTitle = /failed to delete/i;
-        const expectedButtonName = /borrar/i;
-        const button = screen.getByRole("button", {
-          name: expectedButtonName,
-        });
-
-        await user.click(button);
-
-        const failedMessage = await screen.findByText(expectedTitle);
-
-        await expect(failedMessage).toBeInTheDocument();
-      });
     });
   });
 });
