+++ src/artworks/components/ArtworksList/ArtworksList.scss
@@ -1,3 +1,6 @@
 .artworks {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
   width: 100%;
 }
